feat (input): add outline style and success status animation
